refactor(google-auth): moved the sigin button to shared components
